# internal/server/swagger/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/auth/refresh:
        post:
            tags:
                - Auth
            operationId: Auth_RefreshToken
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/goddess.api.v1.RefreshTokenRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.RefreshTokenReply'
    /v1/login/email:
        post:
            tags:
                - Login
            operationId: Login_LoginFromEmail
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/goddess.api.v1.LoginFromEmailRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.LoginReply'
    /v1/login/oauth2:
        post:
            tags:
                - Login
            operationId: Login_LoginFromOAuth2
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/goddess.api.v1.LoginFromOAuth2Request'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.LoginReply'
    /v1/login/phone:
        post:
            tags:
                - Login
            operationId: Login_LoginFromPhone
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/goddess.api.v1.LoginFromPhoneRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.LoginReply'
...
components:
    schemas:
...
        goddess.api.v1.LoginFromEmailRequest:
            type: object
            properties:
                email:
                    type: string
                password:
                    type: string
        goddess.api.v1.LoginFromOAuth2Request:
            type: object
            properties:
                code:
                    type: string
                state:
                    type: string
        goddess.api.v1.LoginFromPhoneRequest:
            type: object
            properties:
                phone:
                    type: string
                password:
                    type: string
        goddess.api.v1.LoginReply:
            type: object
            properties:
                token:
                    type: string
...
        goddess.api.v1.RefreshTokenReply:
            type: object
            properties:
                token:
                    type: string
        goddess.api.v1.RefreshTokenRequest:
            type: object
            properties: {}
...