# pkg/api/v1/login_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v5.29.3
// source: api/v1/login.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationLoginLoginFromEmail = "/goddess.api.v1.Login/LoginFromEmail"
const OperationLoginLoginFromOAuth2 = "/goddess.api.v1.Login/LoginFromOAuth2"
const OperationLoginLoginFromPhone = "/goddess.api.v1.Login/LoginFromPhone"

type LoginHTTPServer interface {
	LoginFromEmail(context.Context, *LoginFromEmailRequest) (*LoginReply, error)
	LoginFromOAuth2(context.Context, *LoginFromOAuth2Request) (*LoginReply, error)
	LoginFromPhone(context.Context, *LoginFromPhoneRequest) (*LoginReply, error)
}

func RegisterLoginHTTPServer(s *http.Server, srv LoginHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/login/email", _Login_LoginFromEmail0_HTTP_Handler(srv))
	r.POST("/v1/login/phone", _Login_LoginFromPhone0_HTTP_Handler(srv))
	r.POST("/v1/login/oauth2", _Login_LoginFromOAuth20_HTTP_Handler(srv))
}

func _Login_LoginFromEmail0_HTTP_Handler(srv LoginHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in LoginFromEmailRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLoginLoginFromEmail)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.LoginFromEmail(ctx, req.(*LoginFromEmailRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*LoginReply)
		return ctx.Result(200, reply)
	}
}

func _Login_LoginFromPhone0_HTTP_Handler(srv LoginHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in LoginFromPhoneRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLoginLoginFromPhone)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.LoginFromPhone(ctx, req.(*LoginFromPhoneRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*LoginReply)
		return ctx.Result(200, reply)
	}
}

func _Login_LoginFromOAuth20_HTTP_Handler(srv LoginHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in LoginFromOAuth2Request
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLoginLoginFromOAuth2)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.LoginFromOAuth2(ctx, req.(*LoginFromOAuth2Request))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*LoginReply)
		return ctx.Result(200, reply)
	}
}

type LoginHTTPClient interface {
	LoginFromEmail(ctx context.Context, req *LoginFromEmailRequest, opts ...http.CallOption) (rsp *LoginReply, err error)
	LoginFromOAuth2(ctx context.Context, req *LoginFromOAuth2Request, opts ...http.CallOption) (rsp *LoginReply, err error)
	LoginFromPhone(ctx context.Context, req *LoginFromPhoneRequest, opts ...http.CallOption) (rsp *LoginReply, err error)
}

type LoginHTTPClientImpl struct {
	cc *http.Client
}

func NewLoginHTTPClient(client *http.Client) LoginHTTPClient {
	return &LoginHTTPClientImpl{client}
}

func (c *LoginHTTPClientImpl) LoginFromEmail(ctx context.Context, in *LoginFromEmailRequest, opts ...http.CallOption) (*LoginReply, error) {
	var out LoginReply
	pattern := "/v1/login/email"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationLoginLoginFromEmail))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *LoginHTTPClientImpl) LoginFromOAuth2(ctx context.Context, in *LoginFromOAuth2Request, opts ...http.CallOption) (*LoginReply, error) {
	var out LoginReply
	pattern := "/v1/login/oauth2"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationLoginLoginFromOAuth2))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *LoginHTTPClientImpl) LoginFromPhone(ctx context.Context, in *LoginFromPhoneRequest, opts ...http.CallOption) (*LoginReply, error) {
	var out LoginReply
	pattern := "/v1/login/phone"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationLoginLoginFromPhone))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}